Move find_free_slots to managers.py
